Jeck - Small fix to r1305. This still leaves card text untranslated, but at least clears the translated text out of memory.
The thing complicating a "100% functionality" fix for this is that all translation changes are destructive- they overwrite the original text. We want this behavior, as it prevents having two sets of strings in memory (the untranslated and translated sets) when card text currently takes hundreds of kb... however, to switch languages we'd want the original text available, which means reloading all the set files. That's not optimal either...

As it is, the player must reset their PSP for translated card text. Again, not optimal. I'm submitting an issue on this topic.