Added 6 cards (card list --> first comment).
All these cards are creatures which have a certain effect when being cast ("played" from owners hand).

Their code profits from the fact that KICKER is not optional in Wagic: This means that whenever you are able to pay Myojin's kicker cost for example (in this case it is {0} --> rate: 100%), it will enter the battlefield with a Divinity counter on it.
- Myojin won't get the Divinity counter
----> when brought into play from the hand with cards like Elvish Piper.
----> when brought into play from any other zone.