changed shuffleTopToBottom method to use STL library to shuffle cards vector. * This resolves a compiler warning found in XCode.
NOTE: This method is not even called in the code anywhere.  The only reference I can find is in the AladdinsLamp.fire() method, but it's commented out.  This method sounds like it should exist, but nothing is using it.  Was this not working before for Aladdin's lamp, and commented out to get it to sort of work?